requirements.txt
botbuilder-dialogs>=4.7.0
botbuilder-ai>=4.7.0
datatypes-date-time>=1.0.0.a2
aiohttp
numpy==1.20.0
pandas==1.1.3
botbuilder-applicationinsights
botbuilder-azure
botbuilder-core
botbuilder-schema
botframework-connector
botbuilder-applicationinsights
botbuilder-integration-applicationinsights-aiohttp